8주차 과제 완성

+++ src/main/java/com/example/demo/service/BoardService.java
@@ -2,6 +2,11 @@ package com.example.demo.service;
 
 import java.util.List;
 
+import com.example.demo.domain.Article;
+import com.example.demo.exception.BoardExceptionType;
+import com.example.demo.exception.CommonExceptionType;
+import com.example.demo.exception.CustomException;
+import com.example.demo.repository.ArticleRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,9 +21,11 @@ import com.example.demo.repository.BoardRepository;
 public class BoardService {
 
     private final BoardRepository boardRepository;
+    private final ArticleRepository articleRepository;
 
-    public BoardService(BoardRepository boardRepository) {
+    public BoardService(BoardRepository boardRepository, ArticleRepository articleRepository) {
         this.boardRepository = boardRepository;
+        this.articleRepository = articleRepository;
     }
 
     public List<BoardResponse> getBoards() {
@@ -29,11 +36,15 @@ public class BoardService {
 
     public BoardResponse getBoardById(Long id) {
         Board board = boardRepository.findById(id);
+        if (board == null) throw new CustomException(BoardExceptionType.NOT_FOUND_BOARD);
+
         return BoardResponse.from(board);
     }
 
     @Transactional
     public BoardResponse createBoard(BoardCreateRequest request) {
+        if (request.name() == null) throw new CustomException(CommonExceptionType.BAD_REQUEST_NULL_VALUE);
+
         Board board = new Board(request.name());
         Board saved = boardRepository.insert(board);
         return BoardResponse.from(saved);
@@ -41,12 +52,17 @@ public class BoardService {
 
     @Transactional
     public void deleteBoard(Long id) {
+        List<Article> articles = articleRepository.findAllByBoardId(id);
+        if (!articles.isEmpty())    throw new CustomException(BoardExceptionType.BOARD_HAS_POSTS);
+
         boardRepository.deleteById(id);
     }
 
     @Transactional
     public BoardResponse update(Long id, BoardUpdateRequest request) {
         Board board = boardRepository.findById(id);
+        if (board == null) throw new CustomException(BoardExceptionType.NOT_FOUND_BOARD);
+
         board.update(request.name());
         Board updated = boardRepository.update(board);
         return BoardResponse.from(updated);
